move tokens stuff to logic folder because in the new mockups their functionality moved to safe page
