+++ src/logic/tokens/store/actions/addTokens.js
@@ -1,7 +1,7 @@
 // @flow
 import { Map } from 'immutable'
 import { createAction } from 'redux-actions'
-import { type Token } from '~/routes/tokens/store/model/token'
+import { type Token } from '~/logic/tokens/store/model/token'
 
 export const ADD_TOKENS = 'ADD_TOKENS'
