+++ src/logic/tokens/store/actions/disableToken.js
@@ -1,6 +1,6 @@
 // @flow
 import { createAction } from 'redux-actions'
-import { type Token } from '~/routes/tokens/store/model/token'
+import { type Token } from '~/logic/tokens/store/model/token'
 
 export const DISABLE_TOKEN = 'DISABLE_TOKEN'
