+++ src/routes/safe/component/Balances/Tokens/actions.js
@@ -1,6 +1,6 @@
 // @flow
-import enableToken from '~/routes/tokens/store/actions/enableToken'
-import disableToken from '~/routes/tokens/store/actions/disableToken'
+import enableToken from '~/logic/tokens/store/actions/enableToken'
+import disableToken from '~/logic/tokens/store/actions/disableToken'
 
 export type Actions = {
   enableToken: typeof enableToken,
