+++ src/routes/safe/component/SendToken/index.jsx
@@ -5,8 +5,8 @@ import { connect } from 'react-redux'
 import Stepper from '~/components/Stepper'
 import { sleep } from '~/utils/timer'
 import { type Safe } from '~/routes/safe/store/model/safe'
-import { getStandardTokenContract } from '~/routes/tokens/store/actions/fetchTokens'
-import { type Token } from '~/routes/tokens/store/model/token'
+import { getStandardTokenContract } from '~/logic/tokens/store/actions/fetchTokens'
+import { type Token } from '~/logic/tokens/store/model/token'
 import { isEther } from '~/utils/tokens'
 import { EMPTY_DATA } from '~/logic/wallets/ethTransactions'
 import { toNative } from '~/logic/wallets/tokens'
