+++ src/routes/safe/container/actions.js
@@ -1,6 +1,6 @@
 // @flow
 import fetchSafe from '~/routes/safe/store/actions/fetchSafe'
-import { fetchTokens } from '~/routes/tokens/store/actions/fetchTokens'
+import { fetchTokens } from '~/logic/tokens/store/actions/fetchTokens'
 
 export type Actions = {
   fetchSafe: typeof fetchSafe,
