+++ src/routes/safe/container/selector.js
@@ -7,8 +7,8 @@ import { type Safe } from '~/routes/safe/store/model/safe'
 import { type Owner } from '~/routes/safe/store/model/owner'
 import { type GlobalState } from '~/store'
 import { sameAddress } from '~/logic/wallets/ethAddresses'
-import { activeTokensSelector, orderedTokenListSelector } from '~/routes/tokens/store/selectors'
-import { type Token } from '~/routes/tokens/store/model/token'
+import { activeTokensSelector, orderedTokenListSelector } from '~/logic/tokens/store/selectors'
+import { type Token } from '~/logic/tokens/store/model/token'
 import { safeParamAddressSelector } from '../store/selectors'
 
 export type SelectorProps = {
