+++ src/store/index.js
@@ -1,11 +1,13 @@
 // @flow
 import { createBrowserHistory } from 'history'
 import { connectRouter, routerMiddleware } from 'connected-react-router'
-import { combineReducers, createStore, applyMiddleware, compose, type Reducer, type Store } from 'redux'
+import {
+  combineReducers, createStore, applyMiddleware, compose, type Reducer, type Store,
+} from 'redux'
 import thunk from 'redux-thunk'
 import provider, { PROVIDER_REDUCER_ID, type State as ProviderState } from '~/logic/wallets/store/reducer/provider'
 import safe, { SAFE_REDUCER_ID, type State as SafeState, safesInitialState } from '~/routes/safe/store/reducer/safe'
-import tokens, { TOKEN_REDUCER_ID, type State as TokensState } from '~/routes/tokens/store/reducer/tokens'
+import tokens, { TOKEN_REDUCER_ID, type State as TokensState } from '~/logic/tokens/store/reducer/tokens'
 import transactions, {
   type State as TransactionsState,
   TRANSACTIONS_REDUCER_ID,
@@ -40,5 +42,4 @@ const initialState = {
 
 export const store: Store<GlobalState> = createStore(reducers, initialState, finalCreateStore)
 
-export const aNewStore = (localState?: Object): Store<GlobalState> =>
-  createStore(reducers, localState, finalCreateStore)
+export const aNewStore = (localState?: Object): Store<GlobalState> => createStore(reducers, localState, finalCreateStore)
