+++ src/test/tokens.dom.enabling.test.js
@@ -3,7 +3,6 @@ import * as TestUtils from 'react-dom/test-utils'
 import { List } from 'immutable'
 import { getWeb3 } from '~/logic/wallets/getWeb3'
 import { type Match } from 'react-router-dom'
-import TokenComponent from '~/routes/tokens/component/Token'
 import Checkbox from '@material-ui/core/Checkbox'
 import { getFirstTokenContract, getSecondTokenContract, addTknTo } from '~/test/utils/tokenMovements'
 import { aNewStore } from '~/store'
@@ -11,10 +10,10 @@ import { aMinedSafe } from '~/test/builder/safe.redux.builder'
 import { travelToTokens } from '~/test/builder/safe.dom.utils'
 import { sleep } from '~/utils/timer'
 import { buildMathPropsFrom } from '~/test/utils/buildReactRouterProps'
-import { tokenListSelector, activeTokensSelector } from '~/routes/tokens/store/selectors'
+import { tokenListSelector, activeTokensSelector } from '~/logic/tokens/store/selectors'
 import { getActiveTokenAddresses } from '~/utils/localStorage/tokens'
 import { enableFirstToken, testToken } from '~/test/builder/tokens.dom.utils'
-import * as fetchTokensModule from '~/routes/tokens/store/actions/fetchTokens'
+import * as fetchTokensModule from '~/logic/tokens/store/actions/fetchTokens'
 import * as enhancedFetchModule from '~/utils/fetch'
 
 describe('DOM > Feature > Enable and disable default tokens', () => {
