+++ src/utils/tokens.js
@@ -2,7 +2,7 @@
 import { List } from 'immutable'
 import logo from '~/assets/icons/icon_etherTokens.svg'
 import { getBalanceInEtherOf } from '~/logic/wallets/getWeb3'
-import { makeToken, type Token } from '~/routes/tokens/store/model/token'
+import { makeToken, type Token } from '~/logic/tokens/store/model/token'
 
 export const ETH_ADDRESS = '0'
 export const isEther = (symbol: string) => symbol === 'ETH'
